fix: Modification of existing file removing newline at EOF
This was due to the convention of having "\n" as the text of the EOF fragment.
Due to FragmentMapBuilder calling trim() when replacing an old fragment,
the line break was removed.
The new convention always uses the empty string as the EOF fragment text,
hence the change to FragmentMapBuilder.exitFile()).
The line break is now added to EOF#gap-before.