fix: Treat void and boolean as primitives
The types `void` and `boolean` are now properly treated as primitives.
Previously, this was problematic because `boolean` was not properly boxed where needed,
and was also compared using `Objects.equals`, whereas `==` would have been sufficient.
`void` is actually not very useful as an attribute type, but was added for consistency.